Refactored POR code in separate dir
