+++ configure.ac
@@ -447,6 +447,7 @@ AC_CONFIG_FILES([
        src/ltsmin-lib/Makefile
        src/mc-lib/Makefile
        src/pins-lib/Makefile
+       src/pins-lib/por/Makefile
        src/util-lib/Makefile
        src/vset-lib/Makefile
        src/spg-lib/Makefile
