+++ src/pins-lib/por/pins2pins-por.c
@@ -10,13 +10,13 @@
 #include <hre/user.h>
 #include <ltsmin-lib/ltsmin-standard.h>
 #include <pins-lib/pins.h>
-#include <pins-lib/pins2pins-por.h>
+#include <pins-lib/por/pins2pins-por.h>
 #include <pins-lib/pins-util.h>
-#include <pins-lib/por-ample.h>
-#include <pins-lib/por-beam.h>
-#include <pins-lib/por-deletion.h>
-#include <pins-lib/por-internal.h>
-#include <pins-lib/por-leap.h>
+#include <pins-lib/por/por-ample.h>
+#include <pins-lib/por/por-beam.h>
+#include <pins-lib/por/por-deletion.h>
+#include <pins-lib/por/por-internal.h>
+#include <pins-lib/por/por-leap.h>
 #include <util-lib/bitmultiset.h>
 #include <util-lib/dfs-stack.h>
 #include <util-lib/util.h>
