+++ src/pins-lib/por/por-internal.h
@@ -2,8 +2,8 @@
 #define POR_INTERNAL
 
 #include <pins-lib/pins.h>
-#include <pins-lib/pins2pins-por.h>
-#include <pins-lib/por-leap.h>
+#include <pins-lib/por/pins2pins-por.h>
+#include <pins-lib/por/por-leap.h>
 #include <util-lib/bitmultiset.h>
 #include <util-lib/util.h>
